+++ .github/workflows/blackhole-demo-tests-impl.yaml
@@ -30,12 +30,14 @@ jobs:
           {
             name: "whisper performance",
             arch: blackhole,
+            runs-on: ["in-service", "${{ inputs.runner-label }}", "pipeline-perf"],
             cmd: pytest models/demos/whisper/demo/demo.py --input-path="models/demos/whisper/demo/dataset/conditional_generation" -k "conditional_generation",
             owner_id: U05RWH3QUPM # Salar Hosseini
           },
           {
             name: "llama3-8b performance",
             arch: blackhole,
+            runs-on: ["in-service", "${{ inputs.runner-label }}", "pipeline-perf"],
             # I think we can get rid of TT_CACHE_PATH here by using a NFS export
             cmd:  HF_MODEL=/localdev/blackhole_demos/huggingface_data/meta-llama/Llama-3.1-8B-Instruct TT_CACHE_PATH=/localdev/blackhole_demos/huggingface_data/meta-llama/Llama-3.1-8B-Instruct pytest models/tt_transformers/demo/simple_text_demo.py -k "performance-ci and not performance-ci-stress-1",
             owner_id: U03PUAKE719 # Miguel Tairum
@@ -43,15 +45,20 @@ jobs:
           { # This should be moved to a BH perf regression pipeline in the future
             name: "unet-shallow performance",
             arch: blackhole,
+            runs-on: ["in-service", "${{ inputs.runner-label }}", "pipeline-perf"],
             cmd: pytest -sv models/experimental/functional_unet/tests/test_unet_perf.py -k "test_unet_trace_perf and not test_unet_trace_perf_multi_device",
             owner_id: U06ECNVR0EN # Evan Smal
+          },
+          { # This test shall be executed on the BH with 20 cores
+            name: "panoptic_deeplab demo",
+            arch: blackhole,
+            runs-on: ["tt-ubuntu-2204-p150b-stable"],
+            cmd: 'TT_METAL_CORE_GRID_OVERRIDE_TODEPRECATE="4, 3" pytest models/experimental/panoptic_deeplab/demo/demo.py::test_panoptic_deeplab_demo',
+            owner_id: U08AWRT6U3E # Boris Janjic
           }
         ]
     name: ${{ matrix.test-group.name }}
-    runs-on:
-      - "in-service"
-      - "${{ inputs.runner-label }}"
-      - "pipeline-perf"
+    runs-on: ${{ matrix.test-group.runs-on }}
     container:
       image: ${{ inputs.docker-image || 'docker-image-unresolved' }}
       env:
@@ -65,7 +72,7 @@ jobs:
         - ${{ github.workspace }}/docker-job:/work # Subdir to workaround https://github.com/actions/runner/issues/691
         - /dev/hugepages-1G:/dev/hugepages-1G
         - "/localdev/blackhole_demos/huggingface_data:/localdev/blackhole_demos/huggingface_data${{ inputs.regenerate-ttnn-cache == true && ':rw' || ':ro' }}"
-      options: "--device /dev/tenstorrent"
+      options: "--device /dev/tenstorrent -e TT_GH_CI_INFRA"
     defaults:
       run:
         shell: bash
